% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/elpR_Rumble_SelectionTable_Restructure_function.R
\name{Rumble_Selection_Table_Restructure}
\alias{Rumble_Selection_Table_Restructure}
\title{Rumble Selection Table Restructure}
\usage{
Rumble_Selection_Table_Restructure(x)
}
\arguments{
\item{HH_selection_tables}{Do not change this value}
}
\value{
Output files
\itemize{
\item updated sound check file with new tab called Rumble Detector Summaries (/elpR/Files/sound_check)
\item Duration of good sounds per site
\item Total detections at lowest score per site
\item Total detections above score threshold per site
\item Total detections above score threshold and on random dates per site (will be blank if random dates are not enabled by user)
\item Raven Pro sound selection tables (/elpR/Files/Selection_Tables/rumble/final) in folders by score
\item Empty selection tables. These are dummy selections tables that have 1 selection per sound file which did not have a detection (elpR/Files/Empty_Tables/rumble) greater than the initial detector score
\item Summary of the number of events per sound file (elpR/Files/num_events/rumble)
\item Plots for the number of detections per day per site by the detector score and the filtered score (elpR/Files/num_events/rumble)
\item Dummy selection tables with selections on sound files that did not have a detection above the filtered score (elpR/Files/zero_days_SSTs/rumble)
}
}
\description{
This function will:
\itemize{
\item read in all selection tables from elephant rumble detector output in the 'Packages/elpR/File/Rumble/raw' folder (required that they are in site-folders)
\item restructure those tables by merging, cleaning, and adding new columns
\item merge the detector tables into site-wise selection tables
\item add 'sound check' info to the selection tables (the 'sound_check' function should have been run on associated sound files BEFORE this step)
\item add the random days designation to the 3 randomly selected days per week (user defined by three_rand_days <- "y" or "n")
\item ignore detections on sound files that were marked with 'y' (bad sound file) in the 'Exclude (y/n)' column of the 'Sound Check' Excel file
\item if min_23hrs <- "y" then ignore detections on sound files that were marked with 'e' (bad sound file for elephant analysis) in the 'Exclude (y/n)' column of the 'Sound Check' Excel file
\item if three_rand_days <- "y" then summarize and produce selections tables for the 3 random days per week
\item save selection tables that are filtered by score per table (in the elpR/File/Rumble/Final folder)
\item count the total number of detections per site and day by score and sound file duration (added to the 'sound check' Excel sheet as a tab, saved in the elpR/Files/sound_check folder)
\item calculates the total number of detections per sound file saved as a .txt file in the elpR/File/Num_events folder
\item plot the total detections per score as .png files (saved in elpR/File/Num_events folder)
\item count the total number of days without detections
\item save a selection table of sound files without detections at threshold score (user defined) for random dates (saved in the elpR/Fileszero_days_SSTs/rumble folder)
}
}
\note{
\itemize{
\item The user needs to update the fields in the script.
\item input detector files require Raven Pro columns:
\itemize{
\item Begin File
\item File Offset (s)
\item Begin Time (s)
\item End Time (s)
\item High Freq (Hz)
\item Low Freq (Hz)
\item Score (can be named something else, but must be indicated by user in script)}
The 'sound check' file is required for this function. The name of the file should be unchanged from when it was created. Use the same variable names between the two scripts for each project, deployment, and disk
This script will only generate selection tables and summaries for sites that are included in the corresponding 'sound check' file.
WARNING: New files will overwrite old files with the same name. Be sure to move files of the same name out of the 'Packages/elpR/File/Rumble/Final' folder before running this script if you don't want to overwrite them
}
}
\author{
Bobbi J. Estabrook \href{mailto:bobbi.estabrook@cornell.edu}{bobbi.estabrook@cornell.edu}
}
