% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_datetime_function.R
\name{selection_datetime}
\alias{selection_datetime}
\title{Extract Date/Time from Selection Table}
\usage{
selection_datetime(filename, file_offset_seconds)
}
\arguments{
\item{filename}{A string containing the filename with format "YYYYMMDD_HHMMSS"}

\item{file_offset_seconds}{Numeric offset in seconds (Raven's File Offset (s))}
}
\value{
POSIXct datetime object in new Begin Clock Time
}
\description{
Extract and modify datetime from Begin File name and File offset (s)
}
