% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpR_sound_check_function.R
\name{sound.check}
\alias{sound.check}
\title{Sound Check}
\usage{
sound.check(x)
}
\arguments{
\item{sound_path}{Do not change this value}
}
\value{
Output file
\itemize{
\item generate an Excel file with sound quality summaries on different tabs in the elpR/Files/sound_check folder
\item "Sounds" Tab Includes A list of each sound file, with:
\item Site name: Site name parsed from sound file name
\item	File Path: computer location of sound file
\item	Current File Name: should follow naming convention described in Part 2 (e.g., Kakum_dep03_8kHz_kk01e_R1426_20230530_095123.wav)
\item	Site Name Length Check: marks if number of characters in site portion of the sound name match expected length (5, including extension)
\item	Current File Start DateTime: Date and time based on current sound file name
\item	Calculated End DateTime: calculated from duration of sound file (for continuous data)
\item	Begin Date: The date the sound file begins on (based on YYYYMMDD in file name)
\item	File Duration (s): Duration of the sound file in seconds
\item	Duration Minutes: Duration of the sound file in minutes
\item	Duration Hours: Duration of the sound file in hours
\item	File Duration Check: Checks if file duration matches that specified by the user in script (fileDurationMin)
\item	Duration (DHMS): Duration of the file in days, hours, minutes and seconds
\item	Total Hours in Date per Site: The number of recording hours in the date of the sound file
\item	File Length Check: Marks sound files with length 0 (0 seconds) as "No Data"
\item File Size (GB): Size of the sound file in gigabytes
\item	SampleRate: Sample rate of the sound file in Hz
\item	Sample_Rate_Check: Checks if sample rate matches expected based on used input from script (sample_rate)
\item	Year Check: If year is 2000 (indicating a Swift clock reset), it is marked as the Wrong Year
\item	Expected File Start Difference minutes: Difference in minutes between the start time of the file to the expected start time of the file based on the duration of the previous file.
\item	Sound Gap Check: marks if the sound file represents a gap (when there > 3 minutes between consecutive sound files for a single site) in recording time or overlaps in time with the timestamp of another sound file for the same site. We do not trust the timestamps on sound with overlap or large gaps.
\item Exclude (y/e): Sounds marked as “y” include those with no data or are <1 minute in duration, and sound time overlap. These sounds will be excluded by the Sound Exclude function in later steps. Sounds marked as “e” indicate that the date for this sound file does not contain a total of 23 hours of sounds, and this file will be excluded from the rumble analysis. This will happen with the Hori-Harm Selection Table Restructure script in later steps.
\item	Sounds <1 minute in duration (gunshots & elephants) will be marked with a ‘y’ in Exclude (y/n) column
\item	Note: The detectors will not run on sounds files of 0 s duration and 0 bytes are in the folders, so will be marked with a ‘y’ in the Exclude (y/e) column by the sound.check script
\item	All sounds after sound overlap should be deleted because we can’t trust the timestamp after that point. You will need to evaluate these situations and manually add a “y” to those sounds as needed.
\item	Sounds with less than 23 hours for a day will be marked with an ‘e’. Since the elephant analysis should only be conducted on full days (>23+ hrs) to capture all hours of the survey equally, these sounds will be excluded from the rumble analysis in later steps.
\item	Notes: Summarized notes on sound problems per file
\item	Deployment: From the user-defined deployment_num value in the script
\item	File Name Length Tab: Number of files with normal name length (25 characters) and non-normal name length per site
\item	File Duration Tab: Number of files with expected duration and other durations per site
\item	Wrong Year Tab: Number of files with the wrong year (e.g., 2000) per site
\item	Sound Gap-Overlap Tab: Number of files with overlap or gap in timestamp relative to next files per site
\item	Duration 0s: Number of files with 0s or no data per site
\item	Sound File Count: Number of sound files per site
\item	Sample_Rate: Number of sound files per sample rate per site
\item	Sound Stats: Summaries of sounds per site
}
}
\description{
This function performs the sound quality check on the sound files and produces a report before detectors are to be run.
This function requires:
\itemize{
\item sound file naming structure must include date and time as YYYYMMDD_HHMMSS
\item site name must be included in sound file name
\item site name must follow specific structure: cc##c, where "c" represents a character value and "#" represents a numeric value
}
}
\note{
\itemize{
\item The user needs to update the fields in the script:
\itemize{
\item User specified sound file location = sound_path
\item deployment name (e.g., nn_202302_feb) = deployment_name
\item deployment number (numeric) = deployment_num
\item disk ID (name of the disk with your sound files) = disk_ID
\item directory for where to move bad sound files to = extra_sounds_folder
\item numeric sample rate (the sample rate of the sound files) = sample_rate
\item file duration (expected duration of each sound file) = fileDurationMin
\item sound file extension type (e.g., ".wav", ".aif") = sound_file_ext
\item a .txt file containing site names with a header of "Site".
\item See existing files in the "elpR/Files/sites" folder for required structure of .txt file
}
}
}
\author{
Bobbi J. Estabrook \href{mailto:bobbi.estabrook@cornell.edu}{bobbi.estabrook@cornell.edu}
}
