% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpR_exclude_sounds_function.R
\name{sound.exclude}
\alias{sound.exclude}
\title{Exclude Bad Sounds}
\usage{
sound.exclude(x, y)
}
\arguments{
\item{sound_path}{Do not change this value}

\item{extra_sounds_folder}{Do not change this value}
}
\value{
Output files
\itemize{
\item summary of the duration of the excluded files and save the info as a new tab in the sound_check xlsx file
\item move the Swift configuration and debug .txt files to the swift_files folder (elpR/Files/swift_files) if the user specifies that those files exist in the sound folders
\item merge the Swift Config files into one .txt file and append that to the sound check sheet (if the user specifies that they want the merge and have the Swift files)
}
}
\description{
This function will move sound files that are marked as 'bad' sounds in the sound check spreadsheet to another folder.
The function will:
\itemize{
\item read the excel table produced by the sound_check script (the sound_check file must exist for this function to work and be in the following directory: elpR/Files/sound_check)
\item move all sounds marked with a 'y' in the "Exclude (y)" column of the sound_check.xlsx file to the excluded sounds folder defined in the script
\item Sound files to exclude include files with less than 1 min of recording and bad sound files (automatically marked as 'y" by the sound.check function). If the user manually marked bad sound files or sounds to exclude with a 'y' in the Exclude (y/n) column, it will remove those sound files.
\item For elephant rumbles, excluded sounds are those with less than 22.95 hours or more than 24 hours (automatically marked by the sound.check function)
\item summarize the duration of the excluded files and save the info as a new tab in the sound_check xlsx file
\item merge the Swift Config files into one .txt file and append that to the sound check sheet (if the user specifies that they want the merge and have the Swift files)
\item this will add the site name to Swift .txt files and move them out of the site folder to the ‘~elpR/Files/swift_files folder if have_SwiftFiles was set to “y”.
}
}
\note{
Run this step AFTER running the sound.check function and BEFORE running the detectors on the sound files
The user should confirm in the sound_check excel file which sounds are marked as 'y' in the Exclude (y/n) column to be sure those sounds should be excluded and to add 'y' to sounds that should be excluded but were not marked with a 'y' automatically by the sound.check function.
DO NOT rename the sound_check Excel file after you run the sound.check function. The scripts in this package often look for the sound_check file by the deployment name, deployment number, and disk number that the user specifies.
The 'sound check' file is required for this function. The name of the file should be unchanged from when it was created. Use the same variable names between the two scripts for each project, deployment, and disk
The user needs to update the fields in the script:
\itemize{
\item User specified sound file location = sound_path
\item deployment name (e.g., nn_202302_feb) = deployment_name
\item deployment number (numeric) = deployment_num
\item disk ID (name of the disk with your sound files) = disk_ID
\item a .txt file containing site names with a header of "Site"
\item set have_SwiftFiles to "y" (lower-case) if you have Swift files in the sound folders or to "n" if you do not
\item set merge_swift_files to "y" if you want the swift files to be merged together into one file (this will not delete the single swift files), or to "n" if not. For this to work, all the swift files must have the same structure (columns names and order)
}
WARNING: New files will overwrite old files with the same name. Be sure to move files of the same name out of the 'Packages/elpR/File/sound_check' folder before running this script if you don't want to overwrite them
}
\author{
Bobbi J. Estabrook \href{mailto:bobbi.estabrook@cornell.edu}{bobbi.estabrook@cornell.edu}
}
